Version 28 for Minecraft 1.18.X for Forge 39.
1) Tree Gloom Effect (leaves turning to dead branches) is reported on
the gloom effect info report command and can be turned off by editing
the configuration file.

2) Fixed a bug that could result in loss of boots about 1 in 27
grimheart treasures.

3) Changed so blocks made of stone and wallblocks won't turn to gravel
in the grim area (some players live there you know!)

4) Added Burnishing Stone.  This will repair items placed in the shield
slot by 5% to 10% of max damage.  Amor will reequip itself after being
repaired.

Example a chest with 384 uses has 47 damage- it is repaired for 19 to 38
points.

Example a sword with 1561 uses has 413 damage- it is repaired for 78 to
156 uses.

5) Burnishing stones can drop in grim citadels and rarely from
*particular* humanoid monster bad guys.



This new functionality should be added to 1.19.1 by next week.